requirements.txt
requests
PyGObject
python-dotenv
sudo apt install python3-gi gir1.2-appindicator3-0.1
